Add ObjectDetector with ORB implementation
ObjectDetector takes a scene and returns a MatchResult. Match result contains a ObjectDetection instance that provides keypoints and descriptors about the match. This is necessary to overlay the match onto the scene if required. ObjectDetection may be too specific to ORBs implementation - I do not have enough openCV context to determine this. If it is we may need to update ObjectDetection to be of a GenericType and the link detectors with components that overlay the match result based on the algo