Add MatDecorator interface with ObjectDetector and no op implementations
MatDecorator provided an interface for components managing inputs such as a live camera feed, or a single static image, and the have a provided Mat decorated.

The ObjectDetecionMatDecorator composes a ObjectDetector - itself an interface that can then have different detectors provided based on the desired algo - which is used to determine if an object if found in the provided scene. If it is we draw a box around it with a screen line